Catch Exceptions thrown by HCL2
HCL2 is not able to parse all IAM Policy Documents, as
some Terraform statements can be quite elaborate. Simply
catch Exceptions thrown by HCL2 and package them as a
Parliament finding.

This prevents the entire analysis to be broken off by
tf-parliament, and the caller can decide to consider if
the error is a problem or not.